AUTO_CLEANUP

+++ main.py
@@ -59,13 +59,14 @@ if __name__ == "__main__":
                     break
         range_start = range_start + relativedelta(months=1, day=1)
 
-    logging.info("Clean Up")
     clean_up_start=time.time()
-    today = datetime.date.today()
-    result = db.get_all_search().where(Search.range_end < today).all()
-    for search in result:
-        db.delete_search(search)
-    measure_clean_up = time.time()-clean_up_start
+    if config.AUTO_CLEANUP:
+        logging.info("Clean Up")
+        today = datetime.date.today()
+        result = db.get_all_search().where(Search.range_end < today).all()
+        for search in result:
+            db.delete_search(search)
+        measure_clean_up = time.time()-clean_up_start
     logging.info(f"Running kiwi time:{measure_kiwi}, running database time:{measure_db}, clean up time:{measure_clean_up}")
     send_stat_mail(db, config.SMTP_TO)
     logging.info("Finished")
